+++ src/game/Light.h
@@ -8,6 +8,6 @@ void Light_(const int no);
 void LightSetContainer(int *const no_p1_ptr);
 int LightGetNo(void);
 struct Vec2f *LightGetPositions(void);
-void LightSetPosition(const int h, const int x, const int y);
+void LightSetPosition(const int h, const float x, const float y);
 struct Colour3f *LightGetColours(void);
 void LightDebug(void);
